refactor: enhance middleware with proper context handling and service communication
- Add gin.Context parameter to auth middleware for proper header propagation
- Update authorization middleware to support inter-service communication
- Improve conditional auth middleware with context-aware request handling
- Enhance smart auth middleware with service client integration
- Standardize middleware patterns across authentication layers

+++ middleware/auth_middleware.go
@@ -20,11 +20,11 @@ type AuthResponse struct {
 }
 
 type AuthData struct {
-	UserID uint `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 }
 
 type Claims struct {
-	UserID uint `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 	jwt.RegisteredClaims
 }
